добавляет удаление обработчика кнопки загрузить
