+++ js/main.js
@@ -1,6 +1,5 @@
 import { data } from './data.js';
 import { renderPictures } from './pictures.js';
 import './user-modal.js';
-import './big-picture.js';
 
 renderPictures(data);
